Add query summary to left pane copy
